+++ test/mock/core/pipeline/mock_pipeline_context.cpp
@@ -523,7 +523,7 @@ void PipelineBase::SendEventToAccessibility(const AccessibilityEvent& accessibil
 
 void PipelineBase::OnActionEvent(const std::string& action) {}
 
-void PipelineBase::SetRootSize(double density, int32_t width, int32_t height) {}
+void PipelineBase::SetRootSize(double density, float width, float height) {}
 
 RefPtr<PipelineBase> PipelineBase::GetCurrentContext()
 {
